requirements.txt
requests
beautifulsoup4
python-dotenv
redis
discord.py